Collision detector finish

+++ src/org/academiadecodigo/Rajaj/Player.java
@@ -32,8 +32,8 @@ public class Player extends GameObject implements KeyboardHandler{
     public void move() {}
 
     public void jump() {
-        int initialHeight = getGridPosition().getHeight();
-        int actualHeight = getGridPosition().getHeight();
+        double initialHeight = getGridPosition().getHeight();
+        double actualHeight = getGridPosition().getHeight();
         while (actualHeight - initialHeight != 170) {
             getGridPosition().moveUp();
             actualHeight += 7;
