+++ src/org/academiadecodigo/Rajaj/grid/position/AbstractGridPosition.java
@@ -22,12 +22,12 @@ public abstract class AbstractGridPosition implements GridPosition {
 
     }
 
-    public int getWidth(){
+    public double getWidth(){
         return width;
     }
 
 
-    public int getHeight(){
+    public double getHeight(){
         return height;
     }
